+++ src/main/scala/edu/ie3/simona/ontology/messages/services/ServiceMessage.scala
@@ -7,20 +7,24 @@
 package edu.ie3.simona.ontology.messages.services
 
 import edu.ie3.simona.agent.participant.ParticipantAgent
-
-import java.util.UUID
 import edu.ie3.simona.api.data.ontology.DataMessageFromExt
 import edu.ie3.simona.ontology.messages.Activation
 import edu.ie3.simona.ontology.messages.services.EvMessage.EvInternal
+import edu.ie3.simona.ontology.messages.services.LoadProfileMessage.LoadProfileMessageInternal
 import edu.ie3.simona.ontology.messages.services.WeatherMessage.WeatherInternal
 import edu.ie3.simona.scheduler.ScheduleLock.ScheduleKey
 import edu.ie3.simona.service.ServiceStateData.InitializeServiceStateData
 import org.apache.pekko.actor.typed.ActorRef
 
+import java.util.UUID
+
 /** Collections of all messages, that are send to and from the different
   * services
   */
-sealed trait ServiceMessage extends EvInternal with WeatherInternal
+sealed trait ServiceMessage
+    extends EvInternal
+    with LoadProfileMessageInternal
+    with WeatherInternal
 
 object ServiceMessage {