+++ src/main/scala/edu/ie3/simona/util/ConfigUtil.scala
@@ -133,6 +133,7 @@ object ConfigUtil {
       ParticipantConfigUtil(
         buildUuidMapping(
           Seq(
+            subConfig.bm.individualConfigs,
             subConfig.load.individualConfigs,
             subConfig.fixedFeedIn.individualConfigs,
             subConfig.pv.individualConfigs,
@@ -142,6 +143,7 @@ object ConfigUtil {
           ).flatten
         ),
         Seq(
+          subConfig.bm.defaultConfig,
           subConfig.load.defaultConfig,
           subConfig.fixedFeedIn.defaultConfig,
           subConfig.pv.defaultConfig,