+++ src/test/scala/edu/ie3/simona/util/ConfigUtilSpec.scala
@@ -131,7 +131,7 @@ class ConfigUtilSpec
 
       inside(actual) { case ParticipantConfigUtil(configs, defaultConfigs) =>
         configs shouldBe Map.empty[UUID, LoadRuntimeConfig]
-        defaultConfigs.size shouldBe 7
+        defaultConfigs.size shouldBe 8
 
         inside(defaultConfigs.get(classOf[LoadRuntimeConfig])) {
           case Some(
@@ -191,7 +191,7 @@ class ConfigUtilSpec
           UUID.fromString("49f250fa-41ff-4434-a083-79c98d260a76")
         )
 
-        defaultConfigs.size shouldBe 7
+        defaultConfigs.size shouldBe 8
         inside(defaultConfigs.get(classOf[LoadRuntimeConfig])) {
           case Some(
                 LoadRuntimeConfig(
@@ -402,7 +402,7 @@ class ConfigUtilSpec
           UUID.fromString("49f250fa-41ff-4434-a083-79c98d260a76")
         )
 
-        defaultConfigs.size shouldBe 7
+        defaultConfigs.size shouldBe 8
         inside(defaultConfigs.get(classOf[FixedFeedInRuntimeConfig])) {
           case Some(
                 FixedFeedInRuntimeConfig(